+++ tests/testthat/test-binless.R
@@ -33,9 +33,9 @@ test_that("cont vpc binless vpcstats are correct", {
   os <- get_os()
 
   if(os == "windows"){
-    location=system.file("extdata/Binless","stats.csv",package="tidyvpc")
+    location <-system.file("extdata/Binless","stats.csv",package="tidyvpc")
   } else {
-    location=system.file("extdata/Binless","stats_l.csv",package="tidyvpc")
+    location <-system.file("extdata/Binless","stats_l.csv",package="tidyvpc")
   }
 
   stats <- fread(location, colClasses = c(qname = "factor"))
@@ -70,9 +70,9 @@ test_that("cont vpc binless stratification vpcstats are correct", {
   os <- get_os()
 
   if(os == "windows"){
-    location=system.file("extdata/Binless","strat_stats.csv",package="tidyvpc")
+    location <-system.file("extdata/Binless","strat_stats.csv",package="tidyvpc")
   } else {
-    location=system.file("extdata/Binless","strat_stats_l.csv",package="tidyvpc")
+    location <-system.file("extdata/Binless","strat_stats_l.csv",package="tidyvpc")
   }
 
   stats <- fread(location, colClasses = c(qname = "factor"))
@@ -83,6 +83,34 @@ test_that("cont vpc binless stratification vpcstats are correct", {
     stats,
     tolerance = 0.001
   )
+})
+
+test_that("cont vpc predcorrect binless vpcstats are correct", {
+  skip_on_cran()
+
+  obs_data <- obs_data[MDV == 0]
+  sim_data <- sim_data[MDV == 0 ]
+  sim_data <- sim_data[REP %% 2 == 1]
+  obs_data$PRED <- sim_data[REP == 1, PRED]
+  vpc <- observed(obs_data, x=TIME, y=DV)
+  vpc <- simulated(vpc, sim_data, y=DV)
+  vpc <- predcorrect(vpc, pred = PRED)
+  vpc <- binless(vpc)
+  vpc <- suppressWarnings(vpcstats(vpc))
+
+  os <- get_os()
+
+  if(os == "windows"){
+    location <-system.file("extdata/Binless","predcor_stats.csv",package="tidyvpc")
+  } else {
+    location <-system.file("extdata/Binless","predcor_stats_l.csv",package="tidyvpc")
+  }
+
+  stats <- fread(location, colClasses = c(qname = "factor"))
+
+  expect_equal(vpc$stats, 
+               stats, 
+               tolerance = 0.003)
 })
 
 test_that("cont vpc binless censoring vpcstats are correct", {
@@ -120,7 +148,7 @@ test_that("cat vpc binless vpcstats are correct", {
   vpc <- binless(vpc)
   vpc <- suppressWarnings(vpcstats(vpc, vpc.type = "categorical"))
 
-  location=system.file("extdata/Binless","cat_stats.csv",package="tidyvpc")
+  location <-system.file("extdata/Binless","cat_stats.csv",package="tidyvpc")
 
   stats <- fread(location, colClasses = c(pname = "factor"))
   setkeyv(stats, c("x"))
@@ -140,7 +168,7 @@ test_that("cat vpc binless stratification vpcstats are correct", {
   vpc <- binless(vpc)
   vpc <- suppressWarnings(vpcstats(vpc, vpc.type = "categorical"))
 
-  location=system.file("extdata/Binless","cat_strat_stats.csv",package="tidyvpc")
+  location <-system.file("extdata/Binless","cat_strat_stats.csv",package="tidyvpc")
 
   stats <- fread(location, colClasses = c(pname = "factor"))
   setkeyv(stats, c(names(vpc$strat), "x"))
@@ -150,31 +178,46 @@ test_that("cat vpc binless stratification vpcstats are correct", {
   expect_equal(vpc$stats, stats)
 })
 
-test_that("binless errors are correct", {
-  ## Subest MDV = 0
+
+test_that("binless.tidyvpcobj returns correct errors and warnings", {
   obs_data <- obs_data[MDV == 0]
   sim_data <- sim_data[MDV == 0]
   obs_data$PRED <- sim_data[REP == 1, PRED]
+  vpc <- observed(obs_data, x=TIME, y=DV)
+  vpc <- simulated(vpc, sim_data, y=DV)
 
-  vpc <- observed(obs_data, x = TIME, y = DV )
-  vpc <- simulated(vpc, sim_data, y = DV)
+  # setting optmize = FALSE without specifying lambda or sp
+  expect_error(binless(vpc, optimize = FALSE),
+               regexp = "Set optimize = TRUE if no lambda or sp arguments specified",
+               fixed = TRUE)
+  
+  #loess.ypc argument is deprecated
+  expect_warning(binless(vpc, loess.ypc = TRUE),
+                 regexp = "The loess.ypc argument is deprecated and will be ignored. Usage of `binless()` with `predcorrect()` will now perform LOESS prediction corrected VPC by default.",
+                 fixed = TRUE)
+  
+})
 
-  expect_error(binless(vpc, optimize = FALSE))
-
-  user_lambda <- data.frame(GENDER_F = c(2,4,2), GENDER_M = c(1.9,3,2.25) )
-
+test_that("binless.tidyvpcobj uses supplied lambda and span if optimize = FALSE", {
+  # continuous VPC
+  obs_data <- obs_data[MDV == 0]
+  sim_data <- sim_data[MDV == 0]
+  obs_data$PRED <- sim_data[REP == 1, PRED]
   vpc <- observed(obs_data, x=TIME, y=DV)
   vpc <- simulated(vpc, sim_data, y=DV)
   vpc <- stratify(vpc, ~ GENDER)
-  expect_error(binless(vpc, span = c(.6, .85), loess.ypc = FALSE))
-
-  vpc <- binless(vpc, loess.ypc=TRUE)
-  expect_error(vpcstats(vpc))
-
-  vpc <- predcorrect(vpc, pred=PRED)
-  vpc <- binless(vpc, lambda = user_lambda, loess.ypc = TRUE, span = c(.6, .85))
-  expect_true(is.data.frame(vpc$vpc.method$lambda))
-
+  
+  user_lambda <- data.frame(GENDER_F = c(2,4,2), GENDER_M = c(1.9,3,2.25))
+  
+  vpc <- binless(vpc, lambda = user_lambda, span = c(0.5, 0.8))
+  vpc <- predcorrect(vpc, pred = PRED)
+  
+  expect_true(vpc$vpc.method$loess.ypc)
+  expect_equal(vpc$vpc.method$lambda, user_lambda)
+  expect_equal(vpc$vpc.method$span, c(0.5, 0.8))
+  
   vpc <- suppressWarnings(vpcstats(vpc))
   expect_s3_class(vpc, "tidyvpcobj")
 })
+
+